tests: Don't check for Node's `.status field.nodeInfo.kubeProxyVersion` field
The `.status field.nodeInfo.kubeProxyVersion` is a lie since its initial introduction. The field is set by kubelet which cannot know the kube-proxy version or whether kube-proxy is running at all or not.
The `DisableNodeKubeProxyVersion` feature gate is enabled by default since K8s 1.31. The field is set to empty string in the Node status.
There is no added value in checking this field in the upgrade tests due to the reasons from above.

Ref https://github.com/kubernetes/enhancements/issues/4004