Remove unnecessary logic for the `KubeletCgroupDriverFromCRI` feature gate
Only the feature gate enablement is not enough the new auto-detection flow to be used. The feature depends on a new CRI API that will be present only in containerd 2.0+. Additionally, even if cgroup driver is specified and the new flow auto-detects another one from the CRI, it will ignore the specified cgroup driver and will use the auto-detected one.